ADF 3.18.0

+++ ADFMovieReward.xcframework/ios-arm64_armv7/ADFMovieReward.framework/Headers/ADFmyAppOpenAd.h
@@ -43,7 +43,7 @@ NS_ASSUME_NONNULL_BEGIN
  
   @param timeout 広告準備完了、失敗Callbackが呼ばれるまでのTimeout。1秒から60秒まで設定可能
  */
-- (void)loadWithTimeout:(int)timeout;
+- (void)loadWithTimeout:(float)timeout;
 
 /**
  *  広告が取得済みかどうか
